Set battlemap size

+++ M2TWEOP Code/M2TWEOP library/smallFuncs.h
@@ -12,7 +12,7 @@ namespace smallFuncs
 	void* GetMainStratObject(void*baseObj);
 	NOINLINE EOP_EXPORT void setAncLimit(unsigned char limit);
 	NOINLINE EOP_EXPORT void setEDUUnitsSize(signed short min, signed short max);
-	NOINLINE EOP_EXPORT void setBattlemapSize(signed short x, signed short y);
+	NOINLINE EOP_EXPORT void setBattlemapSize(float x, float y);
 	NOINLINE EOP_EXPORT void setMaxBgSize(unsigned char size);
 	NOINLINE EOP_EXPORT void unlockConsoleCommands();
 	NOINLINE EOP_EXPORT int getBattleCondCode(DWORD condObject);
